pyproject.toml
[project]
name = "banking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.64",
    "dotenv>=0.9.9",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
]
